requirements.txt
disnake
aiosqlite==0.17.0
python-dotenv==0.18.0